Use separate networks in Traefik example
The whole point of using a Docker socket proxy like CetusGuard is to limit access to the Docker socket. The main purpose of Traefik is exposing HTTP/TCP/UDP services to the public. Such services must share a network with Traefik to be discovered and routed to, but this must not be a network with the Docker socket proxy. Only services which actually needs access to the Docker socket must be in such a network, in this case only Traefik.